# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MailBase API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type DomainStatus = (typeof DomainStatus)[keyof typeof DomainStatus];

export const DomainStatus = {
  pending: "pending",
  verified: "verified",
  failed: "failed",
} as const;

export interface DnsRecord {
  record: string;
  name: string;
  type: string;
  ttl: string;
  status: string;
  value: string;
  priority?: number | null;
}

export interface Domain {
  id: number;
  business: string;
  businessLabel: string;
  domain: string;
  fromName: string;
  fromEmail: string;
  resendDomainId?: string | null;
  status: DomainStatus;
  dnsRecords?: DnsRecord[] | null;
  isDefault: boolean;
  createdAt: string;
  verifiedAt?: string | null;
}

export interface CreateDomainInput {
  business: string;
  businessLabel: string;
  domain: string;
  fromName: string;
  fromEmail: string;
  isDefault?: boolean;
}

export interface UpdateDomainInput {
  fromName?: string;
  fromEmail?: string;
  isDefault?: boolean;
  businessLabel?: string;
}

export type ContactMetadata = { [key: string]: unknown } | null;

export interface Contact {
  id: number;
  business: string;
  email: string;
  firstName?: string | null;
  lastName?: string | null;
  tags: string[];
  unsubscribed: boolean;
  unsubscribeToken: string;
  metadata?: ContactMetadata;
  createdAt: string;
}

export type CreateContactInputMetadata = { [key: string]: unknown };

export interface CreateContactInput {
  business: string;
  email: string;
  firstName?: string;
  lastName?: string;
  tags?: string[];
  metadata?: CreateContactInputMetadata;
}

export type UpdateContactInputMetadata = { [key: string]: unknown };

export interface UpdateContactInput {
  email?: string;
  firstName?: string;
  lastName?: string;
  tags?: string[];
  unsubscribed?: boolean;
  metadata?: UpdateContactInputMetadata;
}

export interface BulkImportContactsInput {
  business: string;
  contacts: CreateContactInput[];
}

export interface BulkImportResult {
  imported: number;
  skipped: number;
  errors: string[];
}

export interface ContactList {
  id: number;
  business: string;
  name: string;
  description?: string | null;
  contactCount: number;
  createdAt: string;
}

export interface CreateListInput {
  business: string;
  name: string;
  description?: string;
}

export interface UpdateListInput {
  name?: string;
  description?: string;
}

export interface AddContactsToListInput {
  contactIds: number[];
}

export interface Template {
  id: number;
  business: string;
  name: string;
  subject: string;
  htmlContent: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateTemplateInput {
  business: string;
  name: string;
  subject: string;
  htmlContent: string;
}

export interface UpdateTemplateInput {
  name?: string;
  subject?: string;
  htmlContent?: string;
}

export type CampaignStatus =
  (typeof CampaignStatus)[keyof typeof CampaignStatus];

export const CampaignStatus = {
  draft: "draft",
  sending: "sending",
  sent: "sent",
  failed: "failed",
} as const;

export interface Campaign {
  id: number;
  business: string;
  name: string;
  subject: string;
  domainId: number;
  listId: number;
  templateId?: number | null;
  htmlContent: string;
  status: CampaignStatus;
  sentAt?: string | null;
  recipientCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateCampaignInput {
  business: string;
  name: string;
  subject: string;
  domainId: number;
  listId: number;
  templateId?: number;
  htmlContent: string;
}

export interface UpdateCampaignInput {
  name?: string;
  subject?: string;
  domainId?: number;
  listId?: number;
  templateId?: number;
  htmlContent?: string;
}

export interface DailyStat {
  date: string;
  sent: number;
  opened: number;
  clicked: number;
  bounced: number;
}

export interface AnalyticsOverview {
  totalSent: number;
  totalOpened: number;
  totalClicked: number;
  totalBounced: number;
  openRate: number;
  clickRate: number;
  bounceRate: number;
  dailyStats: DailyStat[];
}

export type EmailEventType =
  (typeof EmailEventType)[keyof typeof EmailEventType];

export const EmailEventType = {
  sent: "sent",
  opened: "opened",
  clicked: "clicked",
  bounced: "bounced",
  unsubscribed: "unsubscribed",
} as const;

export type EmailEventMetadata = { [key: string]: unknown } | null;

export interface EmailEvent {
  id: number;
  campaignId: number;
  contactId: number;
  type: EmailEventType;
  metadata?: EmailEventMetadata;
  createdAt: string;
}

export interface CampaignAnalytics {
  campaignId: number;
  campaignName: string;
  sent: number;
  opened: number;
  clicked: number;
  bounced: number;
  openRate: number;
  clickRate: number;
  bounceRate: number;
  events: EmailEvent[];
}

export type SendTransactionalInputData = { [key: string]: unknown };

export interface SendTransactionalInput {
  business: string;
  type: string;
  toEmail: string;
  data: SendTransactionalInputData;
}

export type TransactionalLogStatus =
  (typeof TransactionalLogStatus)[keyof typeof TransactionalLogStatus];

export const TransactionalLogStatus = {
  sent: "sent",
  failed: "failed",
} as const;

export interface TransactionalLog {
  id: number;
  business: string;
  type: string;
  toEmail: string;
  status: TransactionalLogStatus;
  error?: string | null;
  sentAt: string;
}

export interface MessageResult {
  message: string;
}

export type GetContactsParams = {
  search?: string;
  business?: string;
};

export type GetListsParams = {
  business?: string;
};

export type GetTemplatesParams = {
  business?: string;
};

export type GetCampaignsParams = {
  business?: string;
};

export type GetAnalyticsOverviewParams = {
  business?: string;
  days?: number;
};

export type GetTransactionalLogParams = {
  business?: string;
  limit?: number;
};