# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MailBase API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /domains:
    get:
      operationId: getDomains
      tags: [domains]
      summary: List all domains
      responses:
        "200":
          description: List of domains
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Domain"
    post:
      operationId: createDomain
      tags: [domains]
      summary: Register a new domain
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDomainInput"
      responses:
        "201":
          description: Domain created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Domain"
  /domains/{id}:
    get:
      operationId: getDomain
      tags: [domains]
      summary: Get a domain by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Domain
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Domain"
    put:
      operationId: updateDomain
      tags: [domains]
      summary: Update a domain
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDomainInput"
      responses:
        "200":
          description: Domain updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Domain"
    delete:
      operationId: deleteDomain
      tags: [domains]
      summary: Delete a domain
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Domain deleted
  /domains/{id}/verify:
    post:
      operationId: verifyDomain
      tags: [domains]
      summary: Verify domain DNS records with Resend
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Domain"

  # Contacts
...
  /lists:
    get:
      operationId: getLists
      tags: [lists]
      summary: List all lists
      parameters:
        - name: business
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of lists
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContactList"
    post:
      operationId: createList
      tags: [lists]
      summary: Create a list
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateListInput"
      responses:
        "201":
          description: List created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactList"
  /lists/{id}:
    get:
      operationId: getList
      tags: [lists]
      summary: Get a list by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactList"
    put:
      operationId: updateList
      tags: [lists]
      summary: Update a list
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateListInput"
      responses:
        "200":
          description: List updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactList"
    delete:
      operationId: deleteList
      tags: [lists]
      summary: Delete a list
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: List deleted
  /lists/{id}/contacts:
    get:
      operationId: getListContacts
      tags: [lists]
      summary: Get contacts in a list
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contacts in list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contact"
    post:
      operationId: addContactsToList
      tags: [lists]
      summary: Add contacts to a list
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddContactsToListInput"
      responses:
        "200":
          description: Contacts added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResult"

  # Templates
  /templates:
    get:
      operationId: getTemplates
      tags: [templates]
      summary: List templates
      parameters:
        - name: business
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Template"
    post:
      operationId: createTemplate
      tags: [templates]
      summary: Create a template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTemplateInput"
      responses:
        "201":
          description: Template created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"
  /templates/{id}:
    get:
      operationId: getTemplate
      tags: [templates]
      summary: Get a template by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"
    put:
      operationId: updateTemplate
      tags: [templates]
      summary: Update a template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTemplateInput"
      responses:
        "200":
          description: Template updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"
    delete:
      operationId: deleteTemplate
      tags: [templates]
      summary: Delete a template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Template deleted

  # Campaigns
...
components:
  schemas:
...
    Domain:
      type: object
      properties:
        id:
          type: integer
        business:
          type: string
        businessLabel:
          type: string
        domain:
          type: string
        fromName:
          type: string
        fromEmail:
          type: string
        resendDomainId:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, verified, failed]
        dnsRecords:
          type: array
          nullable: true
          items:
            $ref: "#/components/schemas/DnsRecord"
        isDefault:
          type: boolean
        createdAt:
          type: string
          format: date-time
        verifiedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - business
        - businessLabel
        - domain
        - fromName
        - fromEmail
        - status
        - isDefault
        - createdAt
    DnsRecord:
      type: object
      properties:
        record:
          type: string
        name:
          type: string
        type:
          type: string
        ttl:
          type: string
        status:
          type: string
        value:
          type: string
        priority:
          type: integer
          nullable: true
      required:
        - record
        - name
        - type
        - ttl
        - status
        - value
    CreateDomainInput:
      type: object
      properties:
        business:
          type: string
        businessLabel:
          type: string
        domain:
          type: string
        fromName:
          type: string
        fromEmail:
          type: string
        isDefault:
          type: boolean
      required:
        - business
        - businessLabel
        - domain
        - fromName
        - fromEmail
    UpdateDomainInput:
      type: object
      properties:
        fromName:
          type: string
        fromEmail:
          type: string
        isDefault:
          type: boolean
        businessLabel:
          type: string
    Contact:
      type: object
      properties:
        id:
          type: integer
        business:
          type: string
        email:
          type: string
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        unsubscribed:
          type: boolean
        unsubscribeToken:
          type: string
        metadata:
          type: object
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - business
        - email
        - tags
        - unsubscribed
        - unsubscribeToken
        - createdAt
...
    ContactList:
      type: object
      properties:
        id:
          type: integer
        business:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        contactCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - business
        - name
        - contactCount
        - createdAt
    CreateListInput:
      type: object
      properties:
        business:
          type: string
        name:
          type: string
        description:
          type: string
      required:
        - business
        - name
    UpdateListInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
    AddContactsToListInput:
      type: object
      properties:
        contactIds:
          type: array
          items:
            type: integer
      required:
        - contactIds
    Template:
      type: object
      properties:
        id:
          type: integer
        business:
          type: string
        name:
          type: string
        subject:
          type: string
        htmlContent:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - business
        - name
        - subject
        - htmlContent
        - createdAt
        - updatedAt
    CreateTemplateInput:
      type: object
      properties:
        business:
          type: string
        name:
          type: string
        subject:
          type: string
        htmlContent:
          type: string
      required:
        - business
        - name
        - subject
        - htmlContent
    UpdateTemplateInput:
      type: object
      properties:
        name:
          type: string
        subject:
          type: string
        htmlContent:
          type: string
...
    MessageResult:
      type: object
      properties:
        message:
          type: string
      required:
        - message