# lib/api-zod/src/generated/types/domain.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MailBase API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DnsRecord } from "./dnsRecord";
import type { DomainStatus } from "./domainStatus";

export interface Domain {
  id: number;
  business: string;
  businessLabel: string;
  domain: string;
  fromName: string;
  fromEmail: string;
  resendDomainId?: string | null;
  status: DomainStatus;
  dnsRecords?: DnsRecord[] | null;
  isDefault: boolean;
  createdAt: Date;
  verifiedAt?: Date | null;
}